Refactor contracts to allow for version query
